backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
playwright==1.44.0
pydantic==2.7.1
python-dotenv==1.0.1
httpx==0.27.0
weasyprint==62.3
jinja2==3.1.4
python-multipart==0.0.9
slowapi==0.1.9
aiofiles==23.2.1
